+++ test/CuModMatrix/stripe_mul_test.jl
@@ -17,9 +17,9 @@ function test_stripe_mul()
     A_data[5,5] = 11^7 - 5
     B_data[95,95] = 11^7 - 23
 
-    A = CuModMatrix(A_data,11^7,elem_type=Float64)
-    B = CuModMatrix(B_data,11^7,elem_type=Float64)
-    C = GPUFiniteFieldMatrices.zeros(Float64,100,100,11^7)
+    A = CuModMatrix(A_data,11^7,elem_type=Float32)
+    B = CuModMatrix(B_data,11^7,elem_type=Float32)
+    C = GPUFiniteFieldMatrices.zeros(Float32,100,100,11^7)
     
     GPUFiniteFieldMatrices.stripe_mul!(C,A,B)
 
@@ -36,9 +36,9 @@ function test_stripe_mul()
     A_data[5,5] = 11^7 - 5
     B_data[95,95] = 11^7 - 23
 
-    A = CuModMatrix(A_data,11^7,elem_type=Float64)
-    B = CuModMatrix(B_data,11^7,elem_type=Float64)
-    C = GPUFiniteFieldMatrices.zeros(Float64,s,s,11^7)
+    A = CuModMatrix(A_data,11^7,elem_type=Float32)
+    B = CuModMatrix(B_data,11^7,elem_type=Float32)
+    C = GPUFiniteFieldMatrices.zeros(Float32,s,s,11^7)
     
     println("Multiplying on the GPU...")
     CUDA.@time GPUFiniteFieldMatrices.stripe_mul!(C,A,B)
@@ -57,9 +57,9 @@ function test_stripe_mul()
     A_data[95,95] = 11^7 - 23
     x_data = Base.rand(1:100,100)
 
-    A = CuModMatrix(A_data,11^7,elem_type=Float64)
-    x = CuModVector(x_data,11^7,elem_type=Float64)
-    z = GPUFiniteFieldMatrices.zeros(Float64,100,11^7)
+    A = CuModMatrix(A_data,11^7,elem_type=Float32)
+    x = CuModVector(x_data,11^7,elem_type=Float32)
+    z = GPUFiniteFieldMatrices.zeros(Float32,100,11^7)
     
     GPUFiniteFieldMatrices.stripe_mul!(z,A,x)
 
@@ -77,9 +77,9 @@ function test_stripe_mul()
     A_data[5,5] = 7^8 - 5
     A_data[95,95] = 7^8 - 23
 
-    A = CuModMatrix(A_data,7^8,elem_type=Float64)
-    x = CuModVector(x_data,7^8,elem_type=Float64)
-    z = GPUFiniteFieldMatrices.zeros(Float64,s,7^8)
+    A = CuModMatrix(A_data,7^8,elem_type=Float32)
+    x = CuModVector(x_data,7^8,elem_type=Float32)
+    z = GPUFiniteFieldMatrices.zeros(Float32,s,7^8)
     
     println("Multiplying on the GPU...")
     CUDA.@time GPUFiniteFieldMatrices.stripe_mul!(z,A,x)
